begin integrating into DKG (WIP)
